Out with the old code and in with the new
Found a link to an example that worked, took it and modified it with the code that I had written. Now the two nrf52840 development kits are connecting to each other and setting 8 symbols per bit as over coded phy as connection setting. Next step is to make one advertise periodically and the other sync and receive data from the one that advertises.